requirements.txt
pynput>=1.7.6
Pillow>=10.0.0
pyperclip>=1.8.2
pystray>=0.19.5
winrt-Windows.Media.Ocr
winrt-Windows.Graphics.Imaging
winrt-Windows.Globalization
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
mcp>=1.0.0
tiktoken>=0.7.0
